Bug 7622 inline if statements causing null pointer exceptions
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@7701 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/af/ui/forms/FormDataObjIFace.java
@@ -188,7 +188,7 @@ public interface FormDataObjIFace
     /**
      * @return The parent data object's table id at the time the object is inserted, updated, or deleted.
      */
-    public abstract Short getParentTableId();
+    public abstract Integer getParentTableId();
     
     /**
      * @return The parent data object's primary record id at the time the object is inserted, updated, or deleted.
