+++ src/edu/ku/brc/specify/datamodel/AccessionAttachment.java
@@ -205,9 +205,9 @@ public class AccessionAttachment extends DataModelObjBase implements ObjectAttac
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Accession.getClassTableId();
+        return Accession.getClassTableId();
     }
 
     /* (non-Javadoc)
