+++ src/edu/ku/brc/specify/datamodel/AgentGeography.java
@@ -180,9 +180,9 @@ public class AgentGeography extends DataModelObjBase implements Serializable
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Agent.getClassTableId();
+        return Agent.getClassTableId();
     }
 
     /* (non-Javadoc)
