+++ src/edu/ku/brc/specify/datamodel/AttachmentMetadata.java
@@ -140,9 +140,9 @@ public class AttachmentMetadata extends DataModelObjBase implements java.io.Seri
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Attachment.getClassTableId();
+        return Attachment.getClassTableId();
     }
 
     /* (non-Javadoc)
