+++ src/edu/ku/brc/specify/datamodel/Author.java
@@ -195,9 +195,9 @@ public class Author extends DataModelObjBase implements java.io.Serializable,
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)ReferenceWork.getClassTableId();
+        return ReferenceWork.getClassTableId();
     }
 
     /* (non-Javadoc)
