+++ src/edu/ku/brc/specify/datamodel/CollectingEventAttachment.java
@@ -210,9 +210,9 @@ public class CollectingEventAttachment extends CollectionMember implements Objec
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)CollectingEvent.getClassTableId();
+        return CollectingEvent.getClassTableId();
     }
 
     /* (non-Javadoc)
