+++ src/edu/ku/brc/specify/datamodel/CollectingEventAttribute.java
@@ -616,9 +616,9 @@ public class CollectingEventAttribute extends CollectionMember implements Clonea
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)CollectingEvent.getClassTableId();
+        return CollectingEvent.getClassTableId();
     }
 
     /* (non-Javadoc)
