+++ src/edu/ku/brc/specify/datamodel/CollectionObjectAttr.java
@@ -265,9 +265,9 @@ public class CollectionObjectAttr extends CollectionMember implements AttributeI
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)CollectionObject.getClassTableId();
+        return CollectionObject.getClassTableId();
     }
 
     /* (non-Javadoc)
