+++ src/edu/ku/brc/specify/datamodel/CollectionObjectCitation.java
@@ -180,9 +180,9 @@ public class CollectionObjectCitation extends CollectionMember implements java.i
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)CollectionObject.getClassTableId();
+        return CollectionObject.getClassTableId();
     }
 
     /* (non-Javadoc)
