+++ src/edu/ku/brc/specify/datamodel/ConservDescriptionAttachment.java
@@ -184,9 +184,9 @@ public class ConservDescriptionAttachment extends DataModelObjBase implements Ob
      */
     @Override
     @Transient
-   public Short getParentTableId()
+   public Integer getParentTableId()
     {
-        return (short)ConservDescription.getClassTableId();
+        return ConservDescription.getClassTableId();
     }
 
     /* (non-Javadoc)
