+++ src/edu/ku/brc/specify/datamodel/Discipline.java
@@ -491,9 +491,9 @@ public class Discipline extends UserGroupScope implements java.io.Serializable, 
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Division.getClassTableId();
+        return Division.getClassTableId();
     }
 
     /* (non-Javadoc)
