+++ src/edu/ku/brc/specify/datamodel/FieldNotebookPage.java
@@ -224,9 +224,9 @@ public class FieldNotebookPage extends DisciplineMember
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)FieldNotebookPageSet.getClassTableId();
+        return FieldNotebookPageSet.getClassTableId();
     }
 
     /* (non-Javadoc)
