+++ src/edu/ku/brc/specify/datamodel/FieldNotebookPageSet.java
@@ -281,9 +281,9 @@ public class FieldNotebookPageSet extends DisciplineMember
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)FieldNotebook.getClassTableId();
+        return FieldNotebook.getClassTableId();
     }
 
     /* (non-Javadoc)
