+++ src/edu/ku/brc/specify/datamodel/GeoCoordDetail.java
@@ -376,9 +376,9 @@ public class GeoCoordDetail extends DataModelObjBase
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Locality.getClassTableId();
+        return Locality.getClassTableId();
     }
 
     /* (non-Javadoc)
