+++ src/edu/ku/brc/specify/datamodel/GeographyTreeDef.java
@@ -199,9 +199,9 @@ public class GeographyTreeDef extends BaseTreeDef<Geography, GeographyTreeDef, G
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Discipline.getClassTableId();
+        return Discipline.getClassTableId();
     }
 
     /* (non-Javadoc)
