+++ src/edu/ku/brc/specify/datamodel/Gift.java
@@ -423,9 +423,9 @@ public class Gift extends DisciplineMember implements java.io.Serializable, Prep
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Division.getClassTableId();
+        return Division.getClassTableId();
     }
 
     /* (non-Javadoc)
