+++ src/edu/ku/brc/specify/datamodel/GiftAgent.java
@@ -177,9 +177,9 @@ public class GiftAgent extends CollectionMember implements java.io.Serializable 
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Gift.getClassTableId();
+        return Gift.getClassTableId();
     }
 
     /* (non-Javadoc)
