+++ src/edu/ku/brc/specify/datamodel/LithoStratTreeDef.java
@@ -365,9 +365,9 @@ public class LithoStratTreeDef extends BaseTreeDef<LithoStrat, LithoStratTreeDef
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Discipline.getClassTableId();
+        return Discipline.getClassTableId();
     }
 
     /* (non-Javadoc)
