+++ src/edu/ku/brc/specify/datamodel/LoanReturnPreparation.java
@@ -225,9 +225,9 @@ public class LoanReturnPreparation extends CollectionMember implements java.io.S
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)LoanPreparation.getClassTableId();
+        return LoanPreparation.getClassTableId();
     }
 
     /* (non-Javadoc)
