+++ src/edu/ku/brc/specify/datamodel/Locality.java
@@ -717,9 +717,9 @@ public class Locality extends DisciplineMember implements AttachmentOwnerIFace<L
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Discipline.getClassTableId();
+        return Discipline.getClassTableId();
     }
 
     /* (non-Javadoc)
