+++ src/edu/ku/brc/specify/datamodel/LocalityAttachment.java
@@ -165,9 +165,9 @@ public class LocalityAttachment extends DataModelObjBase implements ObjectAttach
      */
     @Override
     @Transient
-    public Short getParentTableId()
+    public Integer getParentTableId()
     {
-        return (short)Locality.getClassTableId();
+        return Locality.getClassTableId();
     }
 
     /* (non-Javadoc)
